Updated go-pg from v9 to v10
v10 broke the Query interface, all of the were replaced with the Model().Table(tableName) interface.
We can't define a model, because go-pg requires a model struct with an unexported `tableName` field with a `pg:""` tag to define the name of table and the reflect package does not want to allow setting of tags in runtime, and since the package allows dynamic migration table names, that was the best solution I could come up with at the time.